+++ packages/database/src/index.ts
@@ -3,3 +3,4 @@ export { createPrismaIdentityRepositories } from "./repositories";
 export { createPrismaPropertyRepositories } from "./property-repositories";
 export { createPrismaAnalysisRepository } from "./analysis-repositories";
 export { createPrismaReviewTransaction } from "./review-transaction";
+export { createPrismaStoryboardRepositories } from "./storyboard-repositories";